Move grpcTraceBinPropagator under opentelemtry package
